feat: ligue argent

+++ src/main/java/com/deblock/packman/state/SwitchIfNearEnemies.java
@@ -23,7 +23,7 @@ public class SwitchIfNearEnemies implements Behaviours {
 
     @Override
     public Optional<PacMove> nextMove(Game game, List<PacMove> partnerMoves) {
-        List<Position> accessibleCells =  game.accessibleCells(pac.position(), 2, pac.speed());
+        List<Position> accessibleCells =  game.accessibleCells(pac.position(), pac.speed() * 2, 1);
         List<Pac> enemies = game.getEnemiesOn(accessibleCells);
         CGLogger.log("enemies " + enemies);
         if (!enemies.isEmpty()) {
@@ -32,30 +32,49 @@ public class SwitchIfNearEnemies implements Behaviours {
             if (pac.abilityCooldown == 0 && !typeToWin.equals(pac.typeId)) {
                 CGLogger.log("I switch to avoid to be killed");
                 return Optional.of(new Switch(pac, typeToWin));
-            } else if (typeToWin.equals(pac.typeId) && enemy.abilityCooldown > 0) {
+            } else if (typeToWin.equals(pac.typeId) && enemy.abilityCooldown > 0 && enemy.speed() <= pac.speed()) {
                 CGLogger.log("Go to enemies cell to kill it");
                 Path path = new Path();
                 path.add(enemy.position.x, enemy.position.y);
                 return Optional.of(new GoTo(enemy.position, path, pac, "I will kill you"));
-            } else {
+            } else if ((!typeToWin.equals(pac.typeId) && !pac.typeId.equals(enemy.typeId)) || enemy.abilityCooldown == 0){
                 CGLogger.log("I go back. I don't want to die");
-                List<Position> directAccessibleCells =  game.accessibleCells(pac.position(), 1, pac.speed());
+                List<Position> directAccessibleCells =  game.accessibleCells(pac.position(), pac.speed(), 1);
+                CGLogger.log("accessible cells: " + directAccessibleCells);
                 List<Position> safeCells = directAccessibleCells.stream()
-                        .filter(cell -> enemies.stream().noneMatch(e -> e.position.equals(cell)))
+                        .filter(cell -> !haveEnemiesNear(cell, game, enemies, pac.speed()))
                         .collect(Collectors.toList());
                 if (safeCells.size() == 1) {
                     CGLogger.log("found only one safe cell. go to " + safeCells.get(0));
                     return Optional.of(new GoTo(pac, safeCells.get(0), "go back"));
                 } else {
-                    CGLogger.log("found many safe cell. search cell with the much path");
+                    CGLogger.log("found many safe cell. search cell with the much path " + safeCells);
                     Optional<Position> moreSafeCell = safeCells.stream()
-                                        .max(Comparator.comparingInt(cell -> game.getGrid().accessibleCells(cell, 10, 1).size()));
+                                        .max(Comparator.comparingInt(cell -> scoreCell(game, cell, pac.position)));
                     CGLogger.log("found " + moreSafeCell);
                     return moreSafeCell.map(cell -> new GoTo(pac, cell, "go back"));
                 }
-
             }
         }
         return Optional.empty();
+    }
+
+    private int scoreCell(Game game, Position cell, Position pacPosition) {
+        int nbAccessCells = game.getGrid().accessibleCells(cell, 10, 1, pacPosition).size();
+        List<Position> nbAccessibleCellsForPellet = game.getGrid().accessibleCells(cell, 2, 1, pacPosition);
+        int nbPelletAtTwoStep = (int) nbAccessibleCellsForPellet.stream().filter(game::containsPellet).count();
+        CGLogger.log("cell " + cell + ", score " + (nbAccessCells - nbPelletAtTwoStep) + " nbAccessible " + nbAccessCells + " nbPellet " + nbPelletAtTwoStep);
+        return nbAccessCells + nbPelletAtTwoStep;
+    }
+
+    private boolean haveEnemiesNear(Position cell, Game game, List<Pac> enemies, int speed) {
+        List<Position> positions = game.accessibleCells(cell, speed, 1);
+        positions.add(cell);
+        for (Position pos: positions) {
+            if (enemies.stream().anyMatch(e -> e.position.equals(pos))) {
+                return true;
+            }
+        }
+        return false;
     }
 }
